requirements.txt
anthropic>=0.39.0
elevenlabs>=2.0.0
fastapi>=0.110.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
imageio[ffmpeg]>=2.34.0
jinja2>=3.1.0
librosa>=0.10.0
Pillow>=10.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
trafilatura>=1.10.0
uvicorn[standard]>=0.29.0
yt-dlp>=2024.1.0
